chore(vite): fix Vite warning during dev
there's a warning from react/vite in the browser about a hot-reload error that happens when the
entry file "__main__.tsx" hot reloads

fix #217
